ai-agent/requirements.txt
strands-agents==1.6.0
boto3==1.40.21
python-dotenv==1.1.1
